refactored component organization
