+++ components/categories/CategoryDisplay.vue
